Reorganized for standard package setup
