+++ Dockerfile
@@ -3,6 +3,10 @@ FROM python:3.6
 COPY . /app
 WORKDIR /app
 RUN pip install -r requirements.txt
+RUN pip install -e .
+#RUN export FLASK_APP=profile
 EXPOSE 5000
-ENTRYPOINT ["python"]
-CMD ["app.py"]
+#ENTRYPOINT ["flask"]
+#CMD ["run"]
+ENTRYPOINT FLASK_APP=profile flask run --host=0.0.0.0
+
