+++ profile/model.py
@@ -1,4 +1,4 @@
-from database import db
+from profile.database import db
 
 
 class Users(db.Model):
