+++ profile/test.py
@@ -1,4 +1,4 @@
-from app import app
+from profile import app
 import unittest
