+++ setup.py
@@ -1,0 +1,6 @@
+import setuptools
+
+setuptools.setup(
+    name='profile',
+    packages=setuptools.find_packages(),
+)
